refactor: CustomException 추가 및 예외 처리 구조 개선
- CustomException 추가
- 기존 예외 클래스 수정 및 UnauthorizedException 추가
- GlobalExceptionHandler 개선
