+++ src/main/java/com/clover/bookflow/global/exception/CustomException.java
@@ -1,0 +1,16 @@
+package com.clover.bookflow.global.exception;
+
+import com.clover.bookflow.global.errorcode.ErrorCode;
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException {
+
+  private final ErrorCode errorCode;
+
+  public CustomException(ErrorCode errorCode) {
+    super(errorCode.getMessage());
+    this.errorCode = errorCode;
+  }
+
+}
