+++ src/main/java/com/clover/bookflow/global/exception/handler/GlobalExceptionHandler.java
@@ -1,6 +1,7 @@
-package com.clover.bookflow.global.exception;
+package com.clover.bookflow.global.exception.handler;
 
 import com.clover.bookflow.global.errorcode.ErrorCode;
+import com.clover.bookflow.global.exception.CustomException;
 import com.clover.bookflow.global.response.ApiResponse;
 import com.clover.bookflow.global.response.ErrorDetails;
 import java.util.List;
@@ -12,6 +13,13 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
+
+  @ExceptionHandler(CustomException.class)
+  public ResponseEntity<ApiResponse<Void>> handleCustomExceptions(CustomException ex) {
+    ErrorCode errorCode = ex.getErrorCode();
+    return ResponseEntity.status(errorCode.getStatus())
+        .body(ApiResponse.fail(errorCode.getCode(), errorCode.getMessage(), List.of()));
+  }
 
   @ExceptionHandler(MethodArgumentNotValidException.class)
   public ResponseEntity<Object> handleMethodArgumentNotValidException(
@@ -26,26 +34,10 @@ public class GlobalExceptionHandler {
     return ResponseEntity.badRequest().body(ApiResponse.fail("400", "입력값 오류", errors));
   }
 
-  @ExceptionHandler(DuplicateResourceException.class)
-  public ResponseEntity<ApiResponse<Void>> handleDuplicateResourceException(
-      DuplicateResourceException ex) {
-    ErrorCode errorCode = ex.getErrorCode();
-    return ResponseEntity.status(errorCode.getStatus())
-        .body(ApiResponse.fail(errorCode.getCode(), errorCode.getMessage(), List.of()));
-  }
-
-  @ExceptionHandler(BadRequestException.class)
-  public ResponseEntity<ApiResponse<Void>> handleBadRequestException(BadRequestException ex) {
-    ErrorCode errorCode = ex.getErrorCode();
-    return ResponseEntity.status(errorCode.getStatus())
-        .body(ApiResponse.fail(errorCode.getCode(), errorCode.getMessage(), List.of()));
-  }
-
   @ExceptionHandler(Exception.class)
   public ResponseEntity<ApiResponse<Void>> handleException(Exception ex) {
     return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
         .body(ApiResponse.fail("500", "서버 오류", List.of()));
   }
-
 
 }
